serve the latest binary release from github releases
the benefit of this change is that the binaries will no longer have to be
served from a private s3 bucket.

the unfortunate downside of this change is having to change another
variable whenever a release is done, namely the github_artifact_id
which looks like a counter of all github release binaries.