remove custom postgres error codes
Postgres does not support the vendor-specific getErrorCode() API.
Due to this, prior to this change, the PostgresEnvironment used to
map SQL State from SQLException to custom error codes.

This doesn't seem to make a lot of sense, since those error codes
aren't documented anywhere.

Still, in order to not regress on the ability to check for exceptions
in postgres stored procedures, the `execute fixture` functionality
needs to be extended with the ability to verify the SQL state.