Manually tear down test data for StoredProcNoParams
The database object used in this test is persistant (not dropped
at the end of the test) and as the CoreTests suite can now be
executed by database sessions where transactions are not supported
(automatically committed) so we need to ensure that test output
data is torn down. This prevents side effects impacting subsequent
executions of this test or other tests using the same database
object.